--- src/components/calendar/DateList/index.ts
@@ -1,3 +1,0 @@
-import DateList from './DateList';
-
-export default DateList;
