+++ src/components/calendar/SubjectList/index.ts
@@ -1,0 +1,3 @@
+import SubjectList from './SubjectList';
+
+export default SubjectList;
